requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
image==1.5.33
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytz==2022.2.1
qrcode==7.3.1
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
